[Analysis] Improve performance of TestHistory.QueryStats.
The RPC is reading all results for the test on every page which
leads to it scaling poorly for tests with many results. In this CL,
we manually partition the set of results to be queried to avoid
scanning more time partitions than necessary.

Initially, the RPC scans two days of data. It then uses the number of
results obtained from that query to estimate the number days to scan
in the next query to obtain the desired number of results. The results
are typically retrieved in 2-4 queries.

This approach will increase the minimum time for the RPC to return
as there are more Spanner roundtrips, but it will tend to reduce
the long tail duration as less data is scanned for tests with
large histories (to just what is needed for the current page).

The SQL statement is also optimised to remove the need for a Hash aggregate when querying previous IDs. This reduces query time by
about half.

In benchmarks running locally (which have large roundtrip latency to
Spanner!), performance improves from 8-18s on problematic tests
before this change to 2-3 seconds after the change.

BUG=None

Change-Id: I43099437b7a48ba265ff67916ea930ecccaa1dc3
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6918927
Commit-Queue: Patrick Meiring <meiring@google.com>
Reviewed-by: Beining Chen <beining@google.com>