[ResultDB] Add sources model for Android.
This CL introduces a new way of specifying the sources tested
in an invocation, based on submitted android builds. It is understood
that basically all builds are either submitted builds or have a
reference build that is a submitted build.

Notes:

Moving a field into a oneof is binary compatible for binary,
protojson, textjson formats. It also should also not require
changes for clients in python owing to how the protocol buffer
bindings are generated for that language. Go clients will need
to change as the protocol buffer bindings will change but these
are all understood to be under our control.

BUG=b:441369546

Change-Id: I004e8e6a7b5d84f9e7f9b32ac0e2f386a069cf59
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6956709
Commit-Queue: Patrick Meiring <meiring@google.com>
Reviewed-by: Beining Chen <beining@google.com>