[BHN] Add notify_only_gardnered_builders toggle
This CL adds a new parameter for builder health notifiers called
notify_only_gardened_builders. It checks for the existence of
`gardener_rotations` in properties and will ignore the builder
summary if this flag is set to True and the builder is not a gardened
builder.

Bug: 407539622
Change-Id: I95a52b21027bbf4da757517fb3fc02134427f878
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6974200
Reviewed-by: Patrick Meiring <meiring@google.com>
Commit-Queue: Riley Wong <rgw@google.com>