[MILO] Update changepoints UI to reflect updated LUCI Analysis implementation.
To better accomodate the new source model for Android builds, the 99th
percentile changepoint lower bound is changing from inclusive to
exclusive in the backend. This means it will always be specified in
terms of a "real" source position (a source position for which we have
a test result).

BUG=None
TEST=npm run test

Change-Id: I548c173150860a861a377a2fa6daee0cf4a2c41c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6975745
Commit-Queue: Patrick Meiring <meiring@google.com>
Reviewed-by: Mostafa Draz <mdraz@google.com>