[ResultDB] Calculate row sizes more accurately when exporting.
bqartifactexporter is choking on some builds as it is generating
BigQuery requests that are too large. Fix some obvious limitations
in the row size calculation to see if this resolves the problem.

Estimating the actual JSON-encoded size is expensive so I prefer to
start with the binary proto size.

BUG=b:445587521
TEST=Integration tests

Change-Id: I05dad94446371d5f11dc78dc6564dcae92d777f9
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/7005845
Reviewed-by: Tuan Nguyen <nqmtuan@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>