Pre compute columns using ListDevices call
The dimension query tends to be a lot slower than the list devices call
(especially for android devices) so we can use the result of the list
devices call to pre compute a first estimate of the available columns
and start showing the table to the user wile we wait for the
dimensionQuery to finish.
The columns computed this way are not going to be totally accurate since
the devices in the current page might not have all possible labels but
its still better than nothing wile waiting.

Change-Id: I167e6e216fda8e5287854594ce9aebaa3ccebcdd
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/7006150
Commit-Queue: Pietro Scuttari <pietroscutta@google.com>
Commit-Queue: Konrad Polarczyk <polarczyk@google.com>
Reviewed-by: Konrad Polarczyk <polarczyk@google.com>
Auto-Submit: Pietro Scuttari <pietroscutta@google.com>