reauth: improve plugin error logging and error propagation
This CL updates plugin_handler to either log warning/error or to
propagate errors in a few situations, so that we surface all relevant
information to user (on a terminal) to help troubleshooting.

Previously, we early returned on non-successful exit code and
disregarded stdout. This caused us to drop the error information when a plugin exits with non-zero code, and produces a valid
GetAssertionResponse (whose Error attribute indicates the error cause).

Change-Id: I43e811ca1fb12ffe40e0e32b76f374ab3a3d0b4f
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/7014930
Auto-Submit: Jiewei Qian <qjw@chromium.org>
Reviewed-by: Allen Li <ayatane@chromium.org>
Commit-Queue: Allen Li <ayatane@chromium.org>