[tsmon]: Do not log errors on context Canceled
In the fleet console side we receive lots of failed to flush tsmon
metrics (context cancelled). Most likely it is because of when it scales
down, meaning destroys the containers, hence it cancels this process. It
would be nice to suppress this errors unless there is other critical
case that can happen which is worth logging as an error. Anyway if it
happens consistently there is `No successful tsmon flush for...` error
logged as well.

This CL also makes the number of flush workers adjustable via flag.

Bug: b/450195906
Change-Id: I411c81addb664c7bed5cb470536e59772bd5db82
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/7021572
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Vaghinak Vardanyan <vaghinak@google.com>