[ResultDB] Rename state to finalization_state.
In work units and root invocations. This is a breaking API change
albeit hopefully of modest impact to clients as:
- this state is not normally written to directly but via
  Finalize(WorkUnit|RootInvocation). Only rarely is it set via
  Update(WorkUnit|RootInvocation).
- there are no clients reading this data yet.

The motivation for this change is to clear the way for `status` to
be a more meaningful value, e.g. RUNNING, SUCCEEDED, ERRORED, etc.
which users will care about rather than this finalizer state which
is mostly interesting for data exports/downstream ingestions.

This new `status` field is only planned for Work Units, but I am
updating Root Invocations too, for API-level consistency.

Given uploader work is commencing, it is desirable to make this
change as early as possible, hence the proposal to rename the field
in advance of landing the new status field.

A future change will deal with the migration of the Spanner "status"
column to a column with a more appropriate name.

BUG=b:442447678
TEST=Integration tests

Change-Id: I78fd316fb5b58c39d9bb630cadd76cf270b1e896
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/7030740
Reviewed-by: Allen Xie <zhihuixie@google.com>
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>