[ResultDB] Add work unit state.
This introduces tracking of execution state and outcome to
ResultDB, like what is in buildbucket, swarming and AnTS.

State names were inspired by https://google.aip.dev/216, with
the addition of SKIPPED which was inspired by AnTS and makes some
logical sense in the test results domain.

A few differences of note with prior art:
- CANCELLED is spelled two "L"s is used per aip.dev/216 instead of
  the naming in AnTS with one L.
- The QUEUED and PENDING states in AnTS are both folded into
  the PENDING state (taken from aip.dev/216), which can mean either
  waiting for compute resources or for dependencies (e.g. builds)
  to complete.
- The ERROR state in AnTS becomes FAILED.
- Unlike buildbucket, there is no INFRA_FAILED state. I am expecting
  we will add some failure reason/details fields later, this can be
  captured there if it is important.
- The buildbucket SCHEDULED state is renamed PENDING here and
  covers a slightly broader set of cases.
- The AnTS state typically covers the work unit excluding children
  and here it is changed to cover the work unit including children.
  This is intended to promote encapsulation; by looking at the root
  invocation state it should be obvious if there is a problem
  affecting the completeness of the results in the invocation.
  If not, one should be able to trust it has been dealt with by retries.
  It is difficult for aggregation logic in ResultDB or the UI to know
  how to combine child work unit states as it is not always clear if
  children represent retries or a sharding operation or some combination
  thereof.
- Compared to Swarming, many states are folded together (Swarming
  has many final error states and two final cancelled states).

The FinalizeRootInvocation RPC is being removed, and the state and
finalization state of the root invocation will track the root work
unit. It is not clear what value exists from letting a root invocation
finalize and set state independently of the root work unit, but it is
clear what complexity it was/would add to the API.

BUG=b:442447678
TEST=Integration tests

Change-Id: Ieb90f4bf6bc507672af1c9116ead04348a4386af
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/7031659
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>