Fix a bug where Display All Processes is incorrectly toggled
Currently, when the user adds ?package=XYZ to the URL of the
logcat viewer, the logcat viewer will only display lines belonging
to package XYZ. However, when this happens, the Display All
Processes and Display All Named Processes are incorrectly toggled on.
This CL fixes this bug.

Bug: 452101070
Change-Id: Icfcfceee7764da3bce85ea5adc991d5e2eddffe7
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/7068392
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: Martin Kong <martinkong@google.com>