[ResultDB] Resolve DeadlineEnforcer flake.
The test appears to be flaky as sched.Tasks() are sorted by ETA.
The ETA at which two tasks is scheduled depends on whether the computed offset for the tasks (e.g. :22 seconds and :43 seconds past the minute) and whether this is still in the future when appended to the current
minute or not.

Fixing the time used in tests should resolve this issue.

BUG=b:453068522
TEST=Integration tests

Change-Id: I20cc4fbf04e722df21f35d110cfd2f2adf26f24a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/7070726
Commit-Queue: Patrick Meiring <meiring@google.com>
Reviewed-by: Allen Xie <zhihuixie@google.com>