[cipd] Switch to "V4 signing scheme" for GCS signed URLs.
It allows to attach URL query parameters to a signed portion
of the URL (in particular "userProject" in the future).

It is more convoluted than the V2 scheme we were using.
Luckily, cloud.google.com/go/storage implements it in
(almost) the nicest way, so just use their implementation.

One side effect is that we now need to include the query
string (effectively) into the cache key of signed URLs,
since different query strings result in different signed
URLs now.

Tweak the cache setting a tiny but to account for
"more usage".

R=iannucci@chromium.org

Change-Id: I58297418a1c1df3ca2470e732dc5cb8284091d34
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/7077083
Reviewed-by: Robbie Iannucci <iannucci@google.com>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>