[gae] Make sure we do not expose gRPC status of datastore errors.
If a datastore error propagates (perhaps wrapped with %w) all they
way up to the RPC server, this status code may erroneously be
returned to the client.

This is particularly bad for "The referenced transaction has
expired or is no longer valid" errors, which from Datastore API
perspective are INVALID_ARGUMENT, but should be treated as
transient by the server.

R=iannucci@chromium.org
BUG=b/454839438

Change-Id: If80ca21dfc55d4a1b8d2aec06a445efa30d92bfb
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/7082648
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@google.com>