[swarming] Restore accidentally deleted `delete-tasks` queue.
We haven't been deleting old Swarming task for a while now and
there's a huge backlog. Hopefully it will be able to be processed
on its own once the queue is restored.

Slightly tweak the cleanup cron to make it pick up more tasks
in a single run to process the backlog faster.

R=chanli@chromium.org

Change-Id: Ic9e1daf91509a98934f40d5db01371eba1a00934
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/7085046
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>