[package.json] Add incremental lint: runs on modified files
`npm run lint` is really slow on the milo/ui codebase because of the
size of the repo. It would be nice to run only on modified files in
commits since origin/main to reduce the latency of the lint command.

Tested: `npm run lint-inc`
Change-Id: Ie1fbcdf1f4c97fa6d8af2e5b375cf34dcd9c0d12
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/7084004
Reviewed-by: Mostafa Draz <mdraz@google.com>
Commit-Queue: Paul Hobbs <phobbs@google.com>