Split android devices page and chromeos devices page
The 2 pages DO have quite a bit in common, most notably clearing the columns and filters that are not found on the page.
Abstracting some parts of this code makes sense, but I don't want to do too much in this CL, this one is just a raw split.

Bug=b:456144772

Change-Id: I7c7c07a346f1815fcb73d247988e3382747e0715
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/7095759
Reviewed-by: Pietro Scuttari <pietroscutta@google.com>
Commit-Queue: Pietro Scuttari <pietroscutta@google.com>