resultdb: new log highlighting RPC
Remove the old QueryArtifactFailureOnlyLines RPC and add
CompareArtifactLines RPC.  I have not renamed the file to enable
(hopefully) better diffing for the code review.

The old RPC relied on the BigQuery artifact export, meaning:
1. It was slow
2. It did not work on any artifacts that could not be exported (e.g. GCS
   artifacts)

The new RPC is meant to be called in after the QueryRecentPasses RPC in
LUCI Analysis.

This handles large files better than the old implementation, but will
still be slow for them as it needs to rehash the passing files for each
page.  We can fix this with caching in a follow up CL if this is a
problem.

Bug: 454123742
Change-Id: Ica45519c2768918a920cba9a5e6871d63ae29b97
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/7104380
Commit-Queue: Matthew Warton <mwarton@google.com>
Reviewed-by: Patrick Meiring <meiring@google.com>