[lucicfg] Add gitiles-based source implementation.
This should be faster in most cases where we're able to fetch the
target config root via gitiles.

There is a hazard if someone puts the luci config project root at
the base of a massive repo, though.

Note that this doesn't change the access control situation for these
repos; it's just a performance optimization.

R=vadimsh

Bug: 454434947
Change-Id: I202d8f94ebb38d7e0b8d5aafbc4de89cda992e60
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/7110352
Commit-Queue: Robbie Iannucci <iannucci@google.com>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>