Add lucicfg.main_package()
This adds a function that returns a struct containing information about
the main package being executed. At present, the only information it
contains is the path to the main package relative to the root of the
repo that contains it. This will allow custom code to use this path in
outputs without requiring the starlark to explicitly duplicate its file
location.

Bug: 458607662
Change-Id: I7bad7726fef2c6d1666a7895b4e805a5f0e684eb
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/7132606
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>