[ResultDB] Make sources mandatory.
This makes the sources field mandatory on the RootInvocation to
reduce the risk uploaders may accidentally not populate this field.

Given the importance of sources to many analyses (e.g. changepoint
analysis, test exoneration, log comparison) we want to nudge uploaders
to populate this field as best we can.

To compensate for the fact that sources will not be known in all
use cases (e.g. local testing), the requirements of the Source
object are relaxed, to allow the trivial { is_dirty: true } message
to be accepted in case of local sources.

Downstream clients using the sources object, particularly LUCI
Analysis, are updated to handle the relaxed requirements for the
Sources message; such by filtering out results without base_sources
set from the respective analyses.

BUG=b:454117686
TEST=Integration tests

Change-Id: I1888a3156bc97a313bca7c40cdcc00d82678d7f0
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/7144559
Commit-Queue: Patrick Meiring <meiring@google.com>
Reviewed-by: Allen Xie <zhihuixie@google.com>