[cr-buildbucket] Add scheduleBuildOp helper.
It represents an operation of creation of a batch of builds.
It carries the request, request-scoped data (like the bucket
cache) and accumulates errors.

Will be used inside refactored scheduleBuild(...) implementation,
in particular to avoid hard-to-read indirection through index maps.

Note that it doesn't itself implement the logic of scheduling builds.
There will be a function that does that, but it will be relatively
high-level. This simplifies introducing a new code path for
Turbo CI builders.

R=chanli@chromium.org
BUG=b/454194663

Change-Id: I62f1be7986c8d413eafa4d9bd58fc7fd3d9dcb7c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/7170173
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>