[cr-buildbucket] Add scheduleBuildOp.PrefetchBuilders.
This is a replacement for fetchBuilderConfigs which is very
hard to read due to all the index arithmetic and double
pointers.

fetchBuilderConfigs also does 3 things at once, with two
of them unnecessary (it fetches buckets and compute shadow
map, but this has already been done during validation).

R=chanli@chromium.org
BUG=b/454194663

Change-Id: I1a6dd379ce88bdb6c3a9dd5a5a4f6bb40c4556ad
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/7170746
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>