[cr-buildbucket] Add parentsMap.parentInfraForRequest(...).
One of the ScheduleBuild code paths needs BuildInfra of the
parent build (to copy some Buildbucket.Agent fields if
the shadow config has `InheritFromParent`).

It currently refetches it and in a buggy way that doesn't
work fully correctly when parent_build_id is used.

parentsMap already fetched it, just reuse it (this will happen
in a future CL that actually refactors scheduleBuild to use
all new helpers).

R=chanli@chromium.org
BUG=b/454194663

Change-Id: I10805db76f54688921e64547ada11007508e855a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/7170699
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>