[cr-buildbucket] Refactor scheduleBuild.
This introduces a clear place in code where build experiments
are already evaluated, but builds are not dispatched yet. We'll
plug in Turbo CI code path there.

This also removes double-fetching of Bucket and BuildInfra
entities and fixes minor bugs sitting on intersection of
features (not sure which of these combinations are allowed at
all through):
 * Shadow builds assumed the batch has a single parent build,
   which is not correct when parent_build_id is used.
 * Shadow buckets are ignored when template_build_id is used.

R=chanli@chromium.org
BUG=b/454194663

Change-Id: I3472a7999d68555b0a5a58df2656c570bc6eaf7f
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/7171378
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>