[cr-buildbucket] Add Turbo CI fields to Build entity.
`StageAttemptID` indicates if this is a Turbo CI build and
what workplan etc. it belongs to. It is indexed to allow RunStage
to find a build that matches `Stage.created_by` to use it as
a parent buildbucket build.

`StageAttemptToken` will be used to call WriteNodes and also be
propagated to bbagent to allow recipes to call Turbo CI API.

We'll probably want to surface `StageAttemptID` in Build proto
as well at some point (for debugging, BQ export, etc), but this
can be done later.

R=chanli@chromium.org
BUG=b/454194663

Change-Id: I74ba4a530034e984d45334be1d65020b11c7180c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/7171894
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>