[ResultDB] Relax authorisation requirements for producer resource field.
With the new producer_resource being mandatory, an unintended
side-effect has been that all producers must be trusted producers.
This was not the intent. The intent is to ensure the integrity
of links to certain well-known resources (buildbucket builds,
swarming, ATP runs).

Relax authorisation requirements for setting this field.

BUG=None
TEST=Integration tests

Change-Id: I34ea8218efa4ba8951acbd7d6405686d7277ea4b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/7183357
Reviewed-by: Allen Xie <zhihuixie@google.com>
Reviewed-by: Tianle (Jerry) Jin <jjerry@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>