[ResultDB] Update BatchCreateTestResults to dual write to TRv2 table.
Test result uploads to work units will now also write to the new
test results table.

A framework for safely experimenting with test ID sharding
strategies, including four initial strategies, and metrics to
measure the write performance of those strategies, has been
implemented.

There is risk of material CreateTestResults performance regression
as a result of this CL; depending on the sharding strategy this
could manifest from significantly increasing transaction participants,
or from undersharding creating a root invocation-level write bottleneck
equivalent to the throughput of one span server.

As we are in a pre-production phase for the Root Invocations API
it is an excellent opportunity to experiment and learn about Spanner's
performance characteristics with different sharding strategies.

BUG=b:461912636
TEST=Integration tests

Change-Id: I7885a477cb47a57ea9e98d5c181f9c2ad323e8f5
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/7186707
Reviewed-by: Beining Chen <beining@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>