Add sqldb-connection-max-idle-time flag This flags allows to set dbConn.SetConnMaxIdleTime
The reason we need this is that we've been getting random
connection reset by peer errors in the fleet-console and I suspect that
its caused by trying to use a dead connection. By adding a timeout we
will be able to close the connection ourselves before it gets cut off.

Taking a step back it would be nice to be able to modify the db
connection in the application code itself but I am not sure how to allow
for that so I added a flag instead.

Bug: b/462181421
Change-Id: Icf344bb2435c44beaea240bd96aed2815153263e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/7205103
Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>
Reviewed-by: Konrad Polarczyk <polarczyk@google.com>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Auto-Submit: Pietro Scuttari <pietroscutta@google.com>