[ResultDB] Update BatchCreateTestExonerations to dual write to TEV2.
Implements dual write to the new TestExonerationsV2 table for test
exonerations being uploaded to work units.

This CL also improves the tests for BatchCreateTestResults RPC to
validate the realms written to the underlying TRV2 table are correct,
as this information is not visible in the TestResult proto and is
important to the integrity of the ACLing that will occur.

BUG=b:461912636
TEST=Integration tests

Change-Id: I14d58ec1c93665cd146384d4bde3762fd659cf2a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/7202247
Reviewed-by: Beining Chen <beining@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>