[auth-service-ui] Simplify collapsible list and aid easy copy
This CL simplifies the CollapsibleList component by leveraging
React's conditional formatting instead of returning early for
empty lists, which may be empty due to redaction.

This makes it easier to use existing components, such as the
LookupLink for principals.

The Typography component is set to "div" so that copying from
the list doesn't include newlines between elements.

Also changes the expand icons to match existing Milo icons,
with expanded being a down arrow and collapsed signified with
a right arrow. The whole row is actually clickable, so it
shouldn't be an IconButton, but just an icon.

Bug: b/464348398
Change-Id: I4e636ea3257b00512025dd92b465283ea39fc277
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/7206623
Reviewed-by: Rachael Newitt <renewitt@google.com>
Commit-Queue: Anne Redulla <aredulla@google.com>