[ResultDB] Work around Spanner query planner quirks.
When doing a Merge join [1] between TestResultsV2 and
TestExonerationsV2 tables, such as needed to compute test verdicts,
Spanner selects the merge key as the key of the parent table
(RootInvocationShardId) followed by the remaining join columns
sorted lexicographically. (Before this CL, that is: CaseName, CoarseName, FineName, ModuleName, ModuleScheme, ModuleVariantHash).
This results in the Spanner re-sorting the data before the merge which
destroys all benefits from having the table primary key already
sorted in a suitable order and adds to query time. This is
particularly problematic when aggregating large invocations in the
10M test verdict range.

As we do not appear to be able to hint the merge key (order) in SQL,
we rename the columns so that (ModuleName, ModuleScheme,
ModuleVariantHash, T1CoarseName, T2FineName, T3CaseName) are already
lexicographically ordered.

[1]: https://docs.cloud.google.com/spanner/docs/query-execution-operators#merge-join

Applying this change will cause a temporary outage in staging for
uploads to RootInvocations. As dual write is not yet in prod there
should be no impact.

BUG=b:463448408

Change-Id: I336e6481a3bd7855130b1cb9a151aacab89dc4c1
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/7210119
Reviewed-by: Allen Xie <zhihuixie@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>