[TurboCI] Take a much more minimalist approach to rpc helpers.
This ditches the delta scheme for a smallest-useful set of helpers.

The result is that callsites will be a bit more verbose, but it also
means they will mostly be dealing with the raw protos.

Moving from helpers -> proto and from proto -> helpers is also very
simple, and can be done inline at any point.

Error handling now must be done inline with message construction, but
this is also the standard experience with Go (for better or worse).

The number of extra packages and symbols is greatly reduced.

R=chanli, tbain, vadimsh

Bug: 454372345
Change-Id: I8135b76d2eeb49fb28c293176676438d67b8d5dc
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/7214337
Reviewed-by: Tim Bain <tbain@google.com>
Commit-Queue: Robbie Iannucci <iannucci@google.com>