[ResultDB] Remove staging hostname check in test results publisher.
Previously, the test results publisher had a hardcoded check that
prevented it from running on any hostname that did not start with
"staging.". This caused the publisher to silently exit in production
environments (e.g. results.api.luci.app).

This change removes this check to allow the publisher to function in
all environments. It also updates the unit tests to use a non-staging
hostname to verify the fix.

BUG=b:454120970
TEST=Unit tests

Change-Id: I1712699a14413f86bc2e6638b7d3a36570ca7ba7
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/7220372
Commit-Queue: Allen Xie <zhihuixie@google.com>
Reviewed-by: Matthew Warton <mwarton@google.com>