[auth-service-ui] Fix group members list and highlight dupes
This CL fixes the rendering of group members for users that have edit
powers. The item count had been added to the title, which made the
prefix cleaning of items get skipped.

The item count is now done in the GroupFormList (renamed as it's for
a single group). The item values are not cleansed of their prefixes,
but rather, the component used to display, AuthLookupLink, renders
users neatly.

The validation error has now been split into 2 categories:
* invalid values
* duplicate values

A couple of components were also renamed for clarity:
* GroupsFormList -> GroupFormList
* GroupFormNew -> CreateGroupForm

Bug: b/466189649
Change-Id: Idf31c2ab432c49ccafc30604cab1d7c2a72955fe
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/7231366
Commit-Queue: Andrew Mitchell <mitchella@google.com>
Auto-Submit: Anne Redulla <aredulla@google.com>
Reviewed-by: Andrew Mitchell <mitchella@google.com>