[ResultDB] Add initial test aggregations backend.
This includes the following core functionality:
- Single-level queries (in future, these could be strung together
  to query all levels for backend ingestion and pub/sub use cases,
  by starting queries for the deeper level once all results from
  a previous level have been returned).
- Filtering to a given test prefix. This is useful for tree views
  where drill-down occurs on a node at a time.
- Pagination.
- Ordering by ID (natural table order) or based on UI priority.
  UI ordering attracts a performance penalty so has been made
  a special case to avoid paying this cost when natural ordering
  is sufficient (e.g. downstream ingestion / pub/sub use-case).

The design goal of this solution is to aggregate root invocations
with up to ~10 million test results with latency no more than
a few seconds for the most difficult queries, and aggregate with
subsecond latency for anything typical, i.e. <1 million results.

In line with the above, attempt has been made to make the generated
SQL both readable and performant. Test runs of the generated SQL
reveal query plans that look reasonable (and likely to perform
based on previous benchmarking of comparable queries) but this will
need to be verified with a load test. Some tuning of join types
may still be required, in particular for the ModuleSummaries <-->
TestResultAggregates join which currently resolves to an Cross
apply join.

BUG=b:461911366
TEST=Integration tests

Change-Id: I4338cedbddfd220ded15dfa59bdbb42683896ee5
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/7206556
Reviewed-by: Allen Xie <zhihuixie@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>