[ResultDB] Add validation for TestIdentifierPrefix.
Refactor the existing test ID validation methods to handle validation
of partial test identifiers. The old validation becomes a special case
of the more general validation method that also handles partial IDs.

Refine the specification for TestIdentifierPrefix slightly to
standardise the representation of the prefix with level INVOCATION
and provide clients to specify either ModuleVariant or
ModuleVariantHash. The latter is useful for URLs and necessary for
clients with only limited view access to test results, as the
module variant will be masked for them (and only variant_hash will
be available).

BUG=b:461911366
TEST=Integration tests

Change-Id: I4e01adb981954215824a6c39522fc02656e81fe8
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/7231106
Commit-Queue: Patrick Meiring <meiring@google.com>
Reviewed-by: Allen Xie <zhihuixie@google.com>