[ResultDB] Populate BuildDescriptor.url
Add a URL field on BuildDescriptor. The field is populated with a
link to the UI for the referenced build.

This field can be used to avoid embedding knowledge of the build
system link scheme/environments in every UI and simplifies debugging
the meaning of a particular RootInvocation's BuildDescriptor
when using the RPC API.

BUG=None
TEST=Integration tests

Change-Id: I9438ef6f927fdb80adc33ae2241cd399a279ff98
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/7235273
Reviewed-by: Allen Xie <zhihuixie@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>