[MILO] Update AIP-160 filter UI to always quote strings.
The AIP-160 filter-parser is becoming stricter to avoid future
syntactic ambiguities re: field and function references on the
RHS of a filter in future. As part of this change, any strings
on the RHS of a field name must be quoted. Numbers, enums and
durations will be unquoted.

Implicit filters (e.g. loose text not attached to right of a
field name) will continue to allow use of unquoted strings.

BUG=b:467142685

Change-Id: I877a8a5858996ce2c1f434d5e88450e66128a65b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/7259439
Reviewed-by: Mostafa Draz <mdraz@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>