[ResultDB] Add debug log for GCS permission check failure.
This adds a log message when a user is denied permission to reference a GCS bucket. The log includes the bucket, project, and the list of allowed buckets for that project/user context. This is to help debug configuration issues where users appear to have permission but are still being denied.

BUG=b:463239821
TEST=Unit tests

Change-Id: I80326878f0df0f3cbea7d754b1e67769fc159a73
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/7267444
Commit-Queue: Allen Xie <zhihuixie@google.com>
Reviewed-by: Beining Chen <beining@google.com>