Look for .citc sentinel in parent of the repo root
When looking for the repo root, it is assumed the files signifying a
repo are located in the root of the repo. This is true for git repos,
which contain a .git directory in the root of the repo. In the case of
cog, there is a checkout directory containing the .citc directory, but
the repo gets checked out into a subdirectory of the checkout directory.
This means that looking for .citc finds the parent of the root of the
top-level repo. This results in code using the lucicfg.main_package()
function incorrectly getting the checkout directory included in the
path.

This updates the findroot functionality so that the probes are structs
that provide the sentinel filename to look for and a locatedInRepoParent
bool indicating whether the sentinel file is actually located in the
parent of the repo root. For a probe with locatedInRepoParent set to
true, at each directory being probed (skipping the volume root), the
code will look for the sentinel file in the parent directory. If the
sentinel is found in the parent directory then the directory being
probed will be returned.

Bug: 468387469
Change-Id: Ifc3857a6fa80b2bce2633c005b1467738a222b5a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/7264619
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>