[ResultDB] Add TestAggregationPredicate.contains_test_results_filter.
Extend the predicate used by QueryTestAggregations to support
filtering to aggregations that contain particular test results.

The semantics of the AIP-160 filter is defined using a schema
object in the testresultsv2 package.

The new filters are not yet integrated in the QueryTestAggregations RPC,
this will occur in a follow-up.

BUG=b:461911366
TEST=None

Change-Id: I1d5721c1719f7eeb7ef7508d5d2d9080e212adcf
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/7235195
Commit-Queue: Patrick Meiring <meiring@google.com>
Reviewed-by: Allen Xie <zhihuixie@google.com>