[resultdb] reduce the batch size in applyFinalizationUpdates
PublishTestResultsTask and PublishWorkUnitsTask include a list of work units in the payload, the payload has a size limit of 102400 bytes, which can be exceeded with the current batch size of 2000 work units. Each work unit id is at most 100 bytes, so reduce the batch size limit to 1000.

Change-Id: I6a74ae3ffba8837ad24176b86cbe10bbad008c90
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/7279182
Reviewed-by: Patrick Meiring <meiring@google.com>
Commit-Queue: Beining Chen <beining@google.com>