fleet-console: run autorepair shows undefined for dut id's
This is related to the recent changes regarding the DeviceTable that was
made generic. This is a quick fix. Ideally I guess toolbar should be
made generic as well but that's a bit more work. Should be revisited
later.

Bug: b/472473052
Change-Id: I11517d958e3333e1fd3ac5ddd1ee013220cfead7
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/7319386
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Auto-Submit: Vaghinak Vardanyan <vaghinak@google.com>
Commit-Queue: Tiffany Zhang <zhangtiff@chromium.org>